Fix ticket #1 - v1.0 doesn't display tip info with TUJ GE:
The TUJGE addon only provides median, mean and stddev information for
items, so add support for displaying medians in the tooltips and
approximate the latest market price with the median price if the latest
market price data isn't available.

While here, factor the code to query the TUJ database into a separate
private.GetInfo() function.  This follows more closely the template of
the Auc-Stat-Example2 module from the latest (preview) release of
Auctioneer.